Shim actions/*load-artifact
This repository uses a very small portion of the api from
upload-artifact and download-artifact, all of which are available in
`@v3`. As such it's possible to provide a local action that
conditionally selects one of `v3` or `v4`.

All instances of `upload-artifact@v3` can be switched to `@v4` when it
is available.

When `@v4` isn't available, `@v3` will be used.

The logic for "when `@v4` is available" can be adjusted later.

This shimming can also be reverted when the last version of GHES adds
support for `v4`.